usePortfolioAggregationReducer that returns the context
